Changed file_read and file_write methods' size parameter to size_t

+++ src/file/file_posix.c
@@ -58,20 +58,20 @@ static int file_eof_linux(BD_FILE_H *file)
 }
 #endif
 
-static int64_t file_read_linux(BD_FILE_H *file, uint8_t *buf, int64_t size)
+static int64_t file_read_linux(BD_FILE_H *file, uint8_t *buf, size_t size)
 {
     if (size > 0 && size < BD_MAX_SSIZE) {
-        return (int64_t)fread(buf, 1, (size_t)size, (FILE *)file->internal);
+        return (int64_t)fread(buf, 1, size, (FILE *)file->internal);
     }
 
     BD_DEBUG(DBG_FILE | DBG_CRIT, "Ignoring invalid read of size %"PRId64" (%p)\n", size, (void*)file);
     return 0;
 }
 
-static int64_t file_write_linux(BD_FILE_H *file, const uint8_t *buf, int64_t size)
+static int64_t file_write_linux(BD_FILE_H *file, const uint8_t *buf, size_t size)
 {
     if (size > 0 && size < BD_MAX_SSIZE) {
-        return (int64_t)fwrite(buf, 1, (size_t)size, (FILE *)file->internal);
+        return (int64_t)fwrite(buf, 1, size, (FILE *)file->internal);
     }
 
     BD_DEBUG(DBG_FILE | DBG_CRIT, "Ignoring invalid write of size %"PRId64" (%p)\n", size, (void*)file);
